refactor: set_data action code standalone
